+++ dynamite/config_reader.py
@@ -7,6 +7,7 @@ import math
 import logging
 import importlib
 import yaml
+import numpy as np
 
 from datetime import datetime, timezone
 
@@ -548,7 +549,7 @@ class Configuration(object):
             raise ValueError(text)
         logger.info('System assembled')
         self.validate()
-        # logger.debug(f'System: {self.system}')  # logged as part of parspace
+        logger.debug(f'System: {self.system}')
         logger.debug(f'Settings: {self.settings}')
         logger.info('Configuration validated')
 
@@ -558,7 +559,7 @@ class Configuration(object):
 
         self.parspace = parspace.ParameterSpace(self.system)
         logger.info('Instantiated parameter space')
-        logger.debug(f'Parameter space: {self.parspace}')
+        logger.debug(f'Parameter space: {[p for p in self.parspace]}')
 
         self.all_models = model.AllModels(config=self)
         logger.info('Instantiated AllModels object')
@@ -989,22 +990,25 @@ class Configuration(object):
                     for kin_data in c.kinematic_data:
                         check_gh = (kin_data.type == 'GaussHermite')
                         check_bl = (kin_data.type == 'BayesLOSVD')
-                        if (not check_gh) and (not check_bl):
+                        check_pm = (kin_data.type == 'ProperMotions')
+                        if not (check_gh or check_bl or check_pm):
                             txt = 'VisibleComponent kinematics type must be ' \
-                                  'GaussHermite or BayesLOSVD'
+                                  'GaussHermite, BayesLOSVD, or ProperMotions.'
                             self.logger.error(txt)
                             raise ValueError(txt)
-                        if check_bl:
+                        if check_bl or check_pm:
                             # check weight solver type
                             if ws_type == 'LegacyWeightSolver':
                                 txt = "LegacyWeightSolver can't be used with "\
-                                      "BayesLOSVD - use weight-solver type NNLS"
+                                      "BayesLOSVD nor ProperMotions - " \
+                                      "use weight-solver type NNLS"
                                 self.logger.error(txt)
                                 raise ValueError(txt)
                             # check for compatible chi2 variant
                             if which_chi2 == 'kinmapchi2':
                                 txt = 'kinmapchi2 cannot be used with ' \
-                                      'BayesLOSVD - use chi2 or kinchi2.'
+                                      'BayesLOSVD nor ProperMotions - ' \
+                                      'use chi2 or kinchi2.'
                                 self.logger.error(txt)
                                 raise ValueError(txt)
                         else:  # GaussHermite kinematics
@@ -1012,10 +1016,10 @@ class Configuration(object):
                             _ = kin_data.get_data(
                                 self.settings.weight_solver_settings)
                 else:
-                    self.logger.error('VisibleComponent must have kinematics: '
-                                      'either GaussHermite or BayesLOSVD')
-                    raise ValueError('VisibleComponent must have kinematics: '
-                                     'either GaussHermite or BayesLOSVD')
+                    txt = 'VisibleComponent must have kinematics: ' \
+                          'GaussHermite, BayesLOSVD, and/or ProperMotions.'
+                    self.logger.error(txt)
+                    raise ValueError(txt)
                 if c.population_data and len(c.population_data) > 1:
                     txt = 'VisibleComponent can either have 0 or 1 set(s) ' \
                           f'of population data, not {len(c.population_data)}.'
@@ -1103,8 +1107,12 @@ class Configuration(object):
                 stars = self.system.get_unique_bar_component()
             else:
                 stars = self.system.get_unique_triaxial_visible_component()
-            hist_bins = [k.hist_bins % 2 for k in stars.kinematic_data]
-            if any([h == 0 for h in hist_bins]):
+            hist_bins = [[k.hist_bins] for k in stars.kinematic_data
+                                       if isinstance(k.hist_bins, int)]
+            hist_bins += [list(k.hist_bins) for k in stars.kinematic_data
+                                       if not isinstance(k.hist_bins, int)]
+            hist_bins = np.array([i for h in hist_bins for i in h])
+            if np.any(hist_bins % 2 == 0):
                 all_hist_bins = {k.name: k.hist_bins
                                  for k in stars.kinematic_data}
                 txt = 'Value of hist_bins must be odd for all kinematic ' \