+++ dynamite/model.py
@@ -38,9 +38,13 @@ class AllModels(object):
             stars = config.system.get_unique_bar_component()
         else:
             stars = config.system.get_unique_triaxial_visible_component()
+        self.has_losvd_kins = \
+            len([k for k in stars.kinematic_data
+                 if not isinstance(k, dyn.kinematics.ProperMotions)]) > 0
         self.has_pops = len([p for p in stars.population_data
                              if p.kin_aper is None]) > 0
-        self.has_pms = False  # fill in after implementing proper motions
+        self.has_pms = len([k for k in stars.kinematic_data
+                            if isinstance(k, dyn.kinematics.ProperMotions)])>0
         self.set_filename(config.settings.io_settings['all_models_file'])
         self.make_empty_table()
         self.dynamite_parameters = config.parspace.par_names[:]
@@ -305,9 +309,18 @@ class AllModels(object):
                 and os.path.isfile(d + 'orblibbox.dat.bz2')
         if not check:
             check = os.path.isfile(d + 'orblib_qgrid.dat.bz2') \
-                    and os.path.isfile(d + 'orblib_losvd_hist.dat.bz2') \
-                    and os.path.isfile(d + 'orblibbox_qgrid.dat.bz2') \
+                    and os.path.isfile(d + 'orblibbox_qgrid.dat.bz2')
+        orblib_files_ok = orblib_files_ok and check
+        # check for 'regular' losvd kinematics
+        if self.has_losvd_kins:
+            check = os.path.isfile(d + 'orblib_losvd_hist.dat.bz2') \
                     and os.path.isfile(d + 'orblibbox_losvd_hist.dat.bz2')
+            orblib_files_ok = orblib_files_ok and check
+        # additional files from orblib integration
+        extra_files = ['orblib.dat_orbclass.out', 'mass_radmass.dat',
+                       'mass_qgrid.dat', 'mass_aper.dat',
+                       'orblibbox.dat_orbclass.out']
+        check = all([os.path.isfile(d + f) for f in extra_files])
         orblib_files_ok = orblib_files_ok and check
         # files that need to be there if populations with own apertures exist
         if self.has_pops:
@@ -316,8 +329,9 @@ class AllModels(object):
             orblib_files_ok = orblib_files_ok and check
         # files that need to be there if proper motion data exist
         if self.has_pms:
-            # orblib_files_ok = orblib_files_ok
-            pass
+            check = os.path.isfile(d + 'orblib_pm_hist.dat.bz2') \
+                    and os.path.isfile(d + 'orblibbox_pm_hist.dat.bz2')
+            orblib_files_ok = orblib_files_ok and check
         # set the indicator files in the orblib directory
         ind_files = [pathlib.Path(d + f_name + '_done')
                      for f_name in ('tube', 'box', 'tube_box')]
@@ -690,7 +704,7 @@ class AllModels(object):
         n : int, optional
             How many models to get. If negative, all models except the
             n best models will be returned. The default is 10.
-        which_chi2 : str, optional
+        which_chi2 : str or None, optional
             Which chi2 is used for determining the best models. If None, the
             setting from the configuration file will be used.
             The default is None.