+++ dynamite/orblib.py
@@ -59,6 +59,14 @@ class LegacyOrbitLibrary(OrbitLibrary):
         self.in_dir = config.settings.io_settings['input_directory']
         self.orblibs_in_parallel = \
             config.settings.multiprocessing_settings['orblibs_in_parallel']
+        if self.system.is_bar_disk_system():
+            self.stars = self.system.get_unique_bar_component()
+        else:
+            self.stars = self.system.get_unique_triaxial_visible_component()
+        self.n_hist1d = len([k for k in self.stars.kinematic_data
+                             if not isinstance(k, dyn_kin.ProperMotions)])
+        self.n_hist2d = len([k for k in self.stars.kinematic_data
+                             if isinstance(k, dyn_kin.ProperMotions)])
         if len(config.all_models.table) == 0:
             self.velocity_scaling_factor = 1.0
         else:
@@ -79,24 +87,29 @@ class LegacyOrbitLibrary(OrbitLibrary):
         Creates the following output files in ``output/models/*/datfil/``:
             - begin.dat                     (ics for tube orbits)
             - beginbox.dat                  (ics for box orbits)
-            - orblib.dat.bz2                (zipped tube orbit library)
+            - orblib_qgrid.dat.bz2          (zipped tube orbit library)
+            - orblib_losvd_hist.dat.bz2(*)  (zipped losvd 1d histograms)
+            - orblib_pops.dat.bz2(*)        (zipped populations 0d hists)
+            - orblib_pm_hist.dat.bz2(*)     (zipped proper motions 2d histograms)
             - orblib.dat_orbclass.out       (orbit classification for tube orbs)
-            - orblibbox.dat.bz2             (zipped box orbit library)
+            - orblibbox_qgrid.dat.bz2(*)    (zipped box orbit library)
+            - orblibbox_losvd_hist.dat.bz2(*) (zipped losvd 1d histograms)
+            - orblibbox_pops.dat.bz2(*)     (zipped populations 0d hists)
+            - orblibbox_pm_hist.dat.bz2(*)  (zipped proper motions 2d histograms)
             - orblibbox.dat_orbclass.out    (orbit classification for box orbs)
             - mass_aper.dat                 (MGE masses in apertures)
             - mass_qgrid.dat                (MGE masses in 3D grid)
             - mass_radmass.dat              (MGE masses in radial bins)
             - +8 log and status files
+        (*) only created if there are kinematic datasets with losvd 1d
+            histograms, populations datasets, or proper motion datasets, resp.
 
         """
         # check whether orbit library was calculated already
         if not os.path.isfile(self.mod_dir + 'datfil/tube_box_done'):
             # prepare the fortran input files for orblib
             self.create_fortran_input_orblib(self.mod_dir+'infil/')
-            if self.system.is_bar_disk_system():
-                stars = self.system.get_unique_bar_component()
-            else:
-                stars = self.system.get_unique_triaxial_visible_component()
+            stars = self.stars
             # create the kinematics and populations input files for each
             # kinematic dataset and population dataset with own apertures
             pops = [p for p in stars.population_data if p.kin_aper is None]
@@ -144,10 +157,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
         #---------------------------------------------
         #write parameters_pot.in and parameters_lum.in
         #---------------------------------------------
-        if self.system.is_bar_disk_system():
-            stars = self.system.get_unique_bar_component()
-        else:
-            stars = self.system.get_unique_triaxial_visible_component()
+        stars = self.stars
         bh = self.system.get_component_from_class(physys.Plummer)
         # used to derive the viewing angles
         if self.system.is_bar_disk_system():
@@ -247,7 +257,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
         #---------------------------------
         #write orblib.in and orblibbox.in
         #---------------------------------
-        n_psf_kin = len(stars.kinematic_data)
+        n_psf_kin = len(stars.kinematic_data)  # write data for all kinematics
         psf_pop_idx = [i for i, p in enumerate(stars.population_data)
                        if p.kin_aper is None]  # pops with their own apertures
         def write_orblib_dot_in(box=False):
@@ -278,7 +288,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
             label = '[accuracy]'
             line = f"{self.settings['accuracy']}{tab}{label}\n"
             f.write(line)
-            label = '[number of psfs of the kinematic data + population data]'
+            label = '[number of psfs of the kinematic data + proper motions + population data]'
             line = f"{n_psf_kin+len(psf_pop_idx)}{tab}{label}\n"
             f.write(line)
             for i in range(n_psf_kin):
@@ -313,7 +323,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
             # need to have a psf of their own)
             n_aperture_kin = n_psf_kin
             n_aperture_pop = len(psf_pop_idx)
-            label = '[# of apertures of the kinematic data + population data]'
+            label = '[# of apertures of the kinematic data + proper motion data + population data]'
             line = f'{n_aperture_kin + n_aperture_pop}{tab}{label}\n'
             f.write(line)
             for i in np.arange(n_aperture_kin):
@@ -324,7 +334,8 @@ class LegacyOrbitLibrary(OrbitLibrary):
                 line = f'{i+1}{tab}{label}\n'
                 f.write(line)
                 label = f'[histogram dimension for apert {1+i} ({kin_i.name})]'
-                ap_hist_dim = 1  # 1d losvd histograms
+                ap_hist_dim = 2 if isinstance(kin_i, dyn_kin.ProperMotions) \
+                                else 1
                 f.write(f'{ap_hist_dim}{tab}{label}\n')
             for i0, i in enumerate(psf_pop_idx):  # i0 counts aperture sets
                 pop_i = stars.population_data[i]
@@ -339,11 +350,22 @@ class LegacyOrbitLibrary(OrbitLibrary):
                 ap_hist_dim = 0  # 0d histograms (scalars) for populations data
                 f.write(f'{ap_hist_dim}{tab}{label}\n')
             for kin_i in stars.kinematic_data:
-                label = '[vhist width, center and nbins for kinematics ' \
-                        f'{kin_i.name}]'
-                w, c, b = kin_i.hist_width, kin_i.hist_center, kin_i.hist_bins
-                line = f'{w} {c} {b}{tab[:-2]}{label}\n'
-                f.write(line)
+                txt = '[vhist width, center and nbins for kinematics ' \
+                      f'{kin_i.name}]'
+                if isinstance(kin_i, dyn_kin.ProperMotions):
+                    for i in range(2):
+                        label = f'dim {i + 1} of 2 {txt}'
+                        w = kin_i.hist_width[i]
+                        c = kin_i.hist_center[i]
+                        b = kin_i.hist_bins[i]
+                        line = f'{w} {c} {b} {tab[:-2]}{label}\n'
+                        f.write(line)
+                else:
+                    label = f'dim 1 of 1 {txt}'
+                    w, c = kin_i.hist_width, kin_i.hist_center
+                    b = kin_i.hist_bins
+                    line = f'{w} {c} {b}{tab[:-2]}{label}\n'
+                    f.write(line)
             for i in psf_pop_idx:
                 pop_i = stars.population_data[i]
                 label = '[vhist width, center, nbins are trivial for pops ' \
@@ -379,9 +401,14 @@ class LegacyOrbitLibrary(OrbitLibrary):
                 f_name = f'"{o_file}_pops.dat"'
                 f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
                         '[pops \'0d hist\' file]\n')
-            f_name = f'"{o_file}_losvd_hist.dat"'
-            f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
-                    '[orbit losvd 1d hist file]\n')
+            if self.n_hist1d > 0:
+                f_name = f'"{o_file}_losvd_hist.dat"'
+                f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
+                        '[orbit losvd 1d hist file]\n')
+            if self.n_hist2d > 0:
+                f_name = f'"{o_file}_pm_hist.dat"'
+                f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
+                        '[orbit proper motions 2d hist file]\n')
             f_name = f'"{o_file}.dat_orbclass.out"'
             f.write(f'{f_name}{tab[:-4] if len(f_name) >= 32 else tab[:-3]}'
                     '[orbit classification file]\n')
@@ -583,7 +610,8 @@ class LegacyOrbitLibrary(OrbitLibrary):
         txt_file.write('# now run tube and box orbits in parallel\n')
         txt_file.write('(rm -f datfil/orblib_qgrid.dat.tmp '
                        'datfil/orblib_qgrid.dat '
-                       'datfil/orblib_pops.dat datfil/orblib_losvd_hist.dat\n')
+                       'datfil/orblib_pops.dat datfil/orblib_losvd_hist.dat '
+                       'datfil/orblib_pm_hist.dat\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} < infil/orblib.in '
                         '>> datfil/orblib.log\n')
         txt_file.write('rm -f datfil/mass_qgrid.dat datfil/mass_radmass.dat '
@@ -599,7 +627,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
                            '< infil/triaxmass.in >> datfil/triaxmass.log\n')
             txt_file.write(f'{self.legacy_directory}/triaxmassbin '
                            '< infil/triaxmassbin.in >> datfil/triaxmassbin.log\n')
-        for f in 'qgrid', 'pops', 'losvd_hist':
+        for f in 'qgrid', 'pm_hist', 'pops', 'losvd_hist':
             f_name = 'datfil/orblib_' + f + '.dat'
             txt_file.write(f'test -e {f_name} '
                            f'&& rm -f {f_name}.bz2 && bzip2 -k {f_name}\n')
@@ -608,10 +636,11 @@ class LegacyOrbitLibrary(OrbitLibrary):
         txt_file.write('orblib=$!\n')
         txt_file.write('(rm -f datfil/orblibbox_qgrid.dat.tmp '
                        'datfil/orblibbox_qgrid.dat datfil/orblibbox_pops.dat '
-                       'datfil/orblibbox_losvd_hist.dat\n')
+                       'datfil/orblibbox_losvd_hist.dat '
+                       'datfil/orblibbox_pm_hist.dat\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} '
                        '< infil/orblibbox.in >> datfil/orblibbox.log\n')
-        for f in 'qgrid', 'pops', 'losvd_hist':
+        for f in 'qgrid', 'pm_hist', 'pops', 'losvd_hist':
             f_name = 'datfil/orblibbox_' + f + '.dat'
             txt_file.write(f'test -e {f_name} '
                            f'&& rm -f {f_name}.bz2 && bzip2 -k {f_name}\n')
@@ -650,7 +679,9 @@ class LegacyOrbitLibrary(OrbitLibrary):
                        'datfil/orblib_qgrid.dat.bz2 datfil/orblib_pops.dat '
                        'datfil/orblib_pops.dat.bz2 '
                        'datfil/orblib_losvd_hist.dat '
-                       'datfil/orblib_losvd_hist.dat.bz2\n')
+                        'datfil/orblib_losvd_hist.dat.bz2 '
+                        'datfil/orblib_pm_hist.dat '
+                        'datfil/orblib_pm_hist.dat.bz2\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} < infil/orblib.in '
                        '>> datfil/orblib.log\n')
         txt_file.write('rm -f datfil/mass_qgrid.dat datfil/mass_radmass.dat '
@@ -659,7 +690,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
                        '< infil/triaxmass.in >> datfil/triaxmass.log\n')
         txt_file.write(f'{self.legacy_directory}/triaxmassbin '
                        '< infil/triaxmassbin.in >> datfil/triaxmassbin.log\n')
-        for f in 'qgrid', 'pops', 'losvd_hist':
+        for f in 'qgrid', 'pm_hist', 'pops', 'losvd_hist':
             f_name = 'datfil/orblib_' + f + '.dat'
             txt_file.write(f'test -e {f_name} '
                            f'&& bzip2 -kc {f_name} > {f_name}.staging.bz2 '
@@ -684,10 +715,12 @@ class LegacyOrbitLibrary(OrbitLibrary):
                         'datfil/orblibbox_pops.dat '
                         'datfil/orblibbox_pops.dat.bz2 '
                         'datfil/orblibbox_losvd_hist.dat '
-                        'datfil/orblibbox_losvd_hist.dat.bz2\n')
+                        'datfil/orblibbox_losvd_hist.dat.bz2 '
+                        'datfil/orblibbox_pm_hist.dat '
+                        'datfil/orblibbox_pm_hist.dat.bz2\n')
         txt_file.write(f'{self.legacy_directory}/{orb_prgrm} '
                        '< infil/orblibbox.in >> datfil/orblibbox.log\n')
-        for f in 'qgrid', 'pops', 'losvd_hist':
+        for f in 'qgrid', 'pm_hist', 'pops', 'losvd_hist':
             f_name = 'datfil/orblibbox_' + f + '.dat'
             txt_file.write(
                 f'test -e {f_name} '
@@ -749,9 +782,10 @@ class LegacyOrbitLibrary(OrbitLibrary):
         Read orbit library from file datfil/{fileroot}.dat.bz2'
 
         Depending on the DYNAMITE version, the orbit library will reside in
-        either datfil/{fileroot}.dat.bz2 (legacy behavior) or in two separate
-        files datfil/{fileroot}_qgrid.dat.bz2 and
-        datfil/{fileroot}_losvd_hist.dat.bz2 (new behavior).
+        either datfil/{fileroot}.dat.bz2 (legacy behavior) or in separate
+        files datfil/{fileroot}_qgrid.dat.bz2 and one or both of
+        datfil/{fileroot}_losvd_hist.dat.bz2 and
+        datfil/{fileroot}_pm_hist.dat.bz2 (new behavior).
         If both "legacy" and "new" files exist, default to the new behavior.
         With 'new behavior', populations data (projected masses) may exist in
         datfil/{fileroot}_pops.dat.bz2 and can be read by setting pops=True.
@@ -760,37 +794,35 @@ class LegacyOrbitLibrary(OrbitLibrary):
         ----------
         fileroot : string
             This will probably be either 'orblib' or 'orblibbox'.
-        return_intrinsic_moments: boolean
+        return_intrinsic_moments: boolean, optional
             Whether to return_intrinsic_moments of the orblib.
-        pops: boolean
-            False (the default): only read the orbit library and LOSVD
+            The default is False.
+        pops: boolean, optional
+            False: only read the orbit library and velocity
             histograms, as needed by the weight solvers.
             True: only read the population data's orbit densities.
+            The default is False.
 
         Returns
         -------
         If return_intrinsic_moments is False:
         pops==False will return a tuple of type (list, array) where the
-        orbit library LOSVDs are stored in the list of Histogram objects, and
-        the 3D density of the orbits are stored in the array object.
+        orbit library velocities are stored in the list of Histogram objects,
+        and the 3D density of the orbits are stored in the array object.
         pops==True will return a tuple of type (list, None) where the
         populations' projected masses are in the list of Histogram objects.
 
-        return_intrinsic_moments is True will returns a tuple (array, list)
+        return_intrinsic_moments=True will return a tuple (array, list)
         where the array stores the intrinsic moments of the orblib and the
         list contains the bin edges of the 3D grid.
 
         """
-        if self.system.is_bar_disk_system():
-            stars = self.system.get_unique_bar_component()
-        else:
-            stars = self.system.get_unique_triaxial_visible_component()
+        stars = self.stars
         norb = self.settings['nE'] * self.settings['nI2'] * self.settings['nI3']
         ml = self.parset['ml']
         cur_dir = os.getcwd()
         os.chdir(self.mod_dir)
         check = os.path.isfile(f'datfil/{fileroot}_qgrid.dat.bz2')
-        check = check and os.path.isfile(f'datfil/{fileroot}_losvd_hist.dat.bz2')
         os.chdir(cur_dir)
         legacy_file = False if check else True
         if pops and legacy_file:
@@ -860,25 +892,42 @@ class LegacyOrbitLibrary(OrbitLibrary):
         else:
             density_3D = None
 
-        # next, we check whether we need to read the losvd_hist file: either
-        # pops==False or pops==True and some pops and kins share apertures
+        # next, check if we need to read the losvd_hist and the pm_hist files:
+        # note: legacy_file=True never has 2d histograms
+        # either pops=False or pops=True and some pops and kins share apertures
         pops_unique = [p for p in stars.population_data if p.kin_aper is None]
         if pops and len(pops_unique) < len(stars.population_data):
             kin_aper_for_pops = \
                 [i.kin_aper for i in stars.population_data
                             if i.kin_aper is not None]
         if not pops or len(pops_unique) < len(stars.population_data):
-            if not legacy_file:  # open the losvd_hist file if needed
+            n_kins = len(stars.kinematic_data)
+            hist_widths = [k.hist_width for k in stars.kinematic_data]
+            # UNUSED hist_centers = [k.hist_center for k in stars.kinematic_data]
+            hist_bins = [k.hist_bins for k in stars.kinematic_data]
+            # kinematics k have 1d losvd histograms if type(k.hist_bins)=int and
+            # 2d proper motion hists otherwise (k.hist_bins is a 1d array then)
+            hist_dim = [1 if type(k.hist_bins) is int else 2
+                        for k in stars.kinematic_data]
+            if not legacy_file:  # open losvd_hist and pm_hist files if needed
                 os.chdir(self.mod_dir)
-                orblib_file = f'datfil/{fileroot}_losvd_hist.dat.bz2'
-                tmpfname = f'datfil/{fileroot}_losvd_hist_{ml}.dat'
-                subprocess.run(f'bunzip2 -c {orblib_file} > {tmpfname}',
-                               shell=True)
-                # read the fortran file
-                orblib_in = FortranFile(tmpfname, 'r')
-            # read the losvd histogram data
-            # from histogram_setup_write, lines 1917-1926:
-            _ = orblib_in.read_record(np.int32, np.int32, float)
+                if any(i == 1 for i in hist_dim):
+                    orblib_file = f'datfil/{fileroot}_losvd_hist.dat.bz2'
+                    tmpfname = f'datfil/{fileroot}_losvd_hist_{ml}.dat'
+                    subprocess.run(f'bunzip2 -c {orblib_file} > {tmpfname}',
+                                shell=True)
+                    # read the fortran file
+                    orblib_in = FortranFile(tmpfname, 'r')
+                if any(i == 2 for i in hist_dim):
+                    orblib_file = f'datfil/{fileroot}_pm_hist.dat.bz2'
+                    tmpfname_pm = f'datfil/{fileroot}_pm_hist_{ml}.dat'
+                    subprocess.run(f'bunzip2 -c {orblib_file} > {tmpfname_pm}',
+                                   shell=True)
+                    orblib_in_pm = FortranFile(tmpfname_pm, 'r')
+            # read the losvd and pm histogram data
+            # from histogram_setup_write, lines 1917-1926: (losvd only)
+            if legacy_file or any(i == 1 for i in hist_dim):
+                _ = orblib_in.read_record(np.int32, np.int32, float)
             # tmp = orblib_in.read_record(np.int32, np.int32, float)
             # nconstr = tmp[0][0] # = total number of apertures for ALL kinematics
             # nvhist = tmp[1][0] # = (nvbins-1)/2 for histo of FIRST kinematic set
@@ -887,19 +936,25 @@ class LegacyOrbitLibrary(OrbitLibrary):
             # however, orbits are stored N times where N = number of kinematic sets
             # histogram settings for other N-1 sets may be different from the first
             # these aren't stored in orblib.dat so must read from kinematics objects
-            n_kins = len(stars.kinematic_data)
-            hist_widths = [k.hist_width for k in stars.kinematic_data]
-            # UNUSED hist_centers = [k.hist_center for k in stars.kinematic_data]
-            hist_bins = [k.hist_bins for k in stars.kinematic_data]
-            self.logger.debug('Checking number of velocity bins...')
-            if np.any(np.array(hist_bins) % 2 == 0):
-                error_msg = 'All kinematics need odd number of velocity bins.'
+            self.logger.debug(f'{self.mod_dir}{tmpfname}: '
+                              'checking number of velocity bins...')
+            hist_bins_flat = [[h] for h in hist_bins if isinstance(h, int)]
+            hist_bins_flat += [list(h) for h in hist_bins
+                                       if not isinstance(h, int)]
+            hist_bins_flat = np.array([i for h in hist_bins_flat for i in h])
+            if np.any(hist_bins_flat % 2 == 0):
+                error_msg = f'{self.mod_dir}{tmpfname}: all kinematics ' \
+                            'need odd number of velocity bins.'
                 self.logger.error(error_msg)
-                orblib_in.close()
-                os.remove(tmpfname)
+                if legacy_file or any(i == 1 for i in hist_dim):
+                    orblib_in.close()
+                    os.remove(tmpfname)
+                if any(i == 2 for i in hist_dim):
+                    orblib_in_pm.close()
+                    os.remove(tmpfname_pm)
                 os.chdir(cur_dir)
                 raise ValueError(error_msg)
-            self.logger.debug('...checks ok.')
+            self.logger.debug(f'...{self.mod_dir}{tmpfname}: checks ok.')
             n_apertures = [k.n_spatial_bins for k in stars.kinematic_data]
             # get index linking kinematic set to aperture
             # kin_idx_per_ap[i] = N <--> aperture i is from kinematic set N
@@ -913,8 +968,26 @@ class LegacyOrbitLibrary(OrbitLibrary):
             cum_n_apertures = np.cumsum(n_apertures)
             idx_ap_reset = np.concatenate(([0], cum_n_apertures[:-1]))
             # set up a list of arrays to hold the results
-            tmp = zip(hist_bins,n_apertures)
-            velhist0 = [np.zeros((norb, nv, na)) for (nv,na) in tmp]
+            # velhist0 is a list of length n_kins, each element is
+            # EITHER a 3D array of shape (norb, nv, na) where nv is the number
+            # of 1D velocity bins and na is the number of apertures (=spatial
+            # bins) for the kinematic set
+            # OR a 4D array of shape (norb, nv[0], nv[1], na) where nv[0] and
+            # nv[1] are the number of 2D velocity bins for the proper
+            # motion histograms
+            velhist0 = []
+            for kin_idx, nv in enumerate(hist_bins):
+                if hist_dim[kin_idx] == 1:
+                    velhist0 += [np.zeros((norb, nv, n_apertures[kin_idx]))]
+                elif hist_dim[kin_idx] == 2:
+                    velhist0 += [np.zeros((norb,
+                                           nv[0],
+                                           nv[1],
+                                           n_apertures[kin_idx]))]
+                else:
+                    error_msg = 'Invalid histogram dimension.'
+                    self.logger.error(error_msg)    # should never happen
+                    raise ValueError(error_msg)
             # Next read the histograms themselves.
             for j in range(norb):
                 if legacy_file:  # orbit info is interlaced in the legacy file
@@ -928,38 +1001,81 @@ class LegacyOrbitLibrary(OrbitLibrary):
                                                         quad_light_grid_sizes)
                 for i_ap, kin_idx in enumerate(kin_idx_per_ap):
                     i_ap0 = i_ap - idx_ap_reset[kin_idx]
-                    ivmin, ivmax = orblib_in.read_ints(np.int32)
-                    if ivmin <= ivmax:
-                        nv0 = (hist_bins[kin_idx]-1)/2
-                        # ^--- this is an integer since hist_bins is odd
-                        nv0 = int(nv0)
-                        tmp = orblib_in.read_reals(float)
-                        velhist0[kin_idx][j, ivmin+nv0:ivmax+nv0+1, i_ap0] = tmp
-            orblib_in.close()
-            # remove temporary file
-            os.remove(tmpfname)
+                    if hist_dim[kin_idx] == 1:
+                        ivmin, ivmax = orblib_in.read_ints(np.int32)
+                        if ivmin <= ivmax:
+                            nv0 = (hist_bins[kin_idx] - 1) // 2
+                            # ^--- this is an integer since hist_bins is odd
+                            tmp = orblib_in.read_reals(float)
+                            velhist0[kin_idx][j,
+                                              ivmin+nv0:ivmax+nv0+1,
+                                              i_ap0] = tmp
+                    elif hist_dim[kin_idx] == 2:
+                        ivmin0, ivmin1, ivmax0, ivmax1 = \
+                            orblib_in_pm.read_ints(np.int32)
+                        if ivmin0 <= ivmax0 and ivmin1 <= ivmax1:
+                            nv0 = [(hist_bins[kin_idx][0] - 1) // 2,
+                                   (hist_bins[kin_idx][1] - 1) // 2]
+                            # ^--- these are integers since hist_bins is odd
+                            tmp = orblib_in_pm.read_reals(float)
+                            tmp = tmp.reshape((ivmax0-ivmin0+1,
+                                               ivmax1-ivmin1+1), order='F')
+                            velhist0[kin_idx][j,
+                                              ivmin0+nv0[0]:ivmax0+nv0[0]+1,
+                                              ivmin1+nv0[1]:ivmax1+nv0[1]+1,
+                                              i_ap0] = tmp
+                    else:
+                        error_msg = 'Invalid histogram dimension.'
+                        self.logger.error(error_msg)    # should never happen
+                        raise ValueError(error_msg)
+
+            if legacy_file or any(i == 1 for i in hist_dim):
+                orblib_in.close()
+                os.remove(tmpfname)
+            if any(i == 2 for i in hist_dim):
+                orblib_in_pm.close()
+                os.remove(tmpfname_pm)
             os.chdir(cur_dir)
             if return_intrinsic_moments:
                 return intrinsic_moms, intrinsic_grid  #######################
             else:
                 velhists = []
-                for i, velhist in enumerate(velhist0):
-                    if pops and i not in kin_aper_for_pops:
+                for kin_idx, velhist in enumerate(velhist0):
+                    if pops and kin_idx not in kin_aper_for_pops:
                         # if pops and this kinematic set is not used by pops
                         continue
-                    # if pops is False or this kinematic set is used by pops
-                    width0 = hist_widths[i]
-                    bins0 = hist_bins[i]
-                    idx_center = (bins0-1)/2 # integer since hist_bins is odd
-                    idx_center = int(idx_center)
-                    dvhist0 = width0/bins0
-                    vedg = np.arange(bins0+1) * dvhist0
-                    v = (vedg[1:] + vedg[:-1])/2.
-                    v_cent = v[idx_center]
-                    vedg -= v_cent
-                    vvv = dyn_kin.Histogram(xedg=vedg, y=velhist)
-                    velhists += [vvv]
-        else:  # if pops and len(pops_unique) == len(stars.population_data)
+                    # if pops is False or this kinematic set is not used by pops
+                    if hist_dim[kin_idx] == 1:
+                        width = hist_widths[kin_idx]
+                        bins = hist_bins[kin_idx]
+                        idx_center = (bins-1)/2 # integer since hist_bins is odd
+                        idx_center = int(idx_center)
+                        dvhist0 = width / bins
+                        vedg = np.arange(bins + 1) * dvhist0
+                        v = (vedg[1:] + vedg[:-1])/2.
+                        v_cent = v[idx_center]
+                        vedg -= v_cent
+                        vvv = dyn_kin.Histogram(xedg=vedg, y=velhist)
+                        velhists += [vvv]
+                    elif hist_dim[kin_idx] == 2:
+                        width = np.array(hist_widths[kin_idx])
+                        bins = np.array(hist_bins[kin_idx], dtype=int)
+                        # dvhist = width / bins
+                        vedg = (np.linspace(start=-width[0] / 2,
+                                            stop=width[0] / 2,
+                                            num=bins[0] + 1,
+                                            endpoint=True),
+                                np.linspace(start=-width[1] / 2,
+                                            stop=width[1] / 2,
+                                            num=bins[1] + 1,
+                                            endpoint=True))
+                        vvv = dyn_kin.Histogram2D(xedg=vedg, y=velhist)
+                        velhists += [vvv]
+                    else:
+                        error_msg = 'Invalid histogram dimension.'
+                        self.logger.error(error_msg)    # should never happen
+                        raise ValueError(error_msg)
+        else:
             velhists = []
 
         if pops and len(pops_unique) > 0:  # read remaining population data
@@ -1001,25 +1117,40 @@ class LegacyOrbitLibrary(OrbitLibrary):
 
         Parameters
         ----------
-        orblib : ``dyn.kinematics.Histogram``
+        orblib : ``dyn.kinematics.Histogram`` or ``dyn.kinematics.Histogram2D``
 
         Returns
         -------
-        ``dyn.kinematics.Histogram``
+        ``dyn.kinematics.Histogram`` or ``dyn.kinematics.Histogram2D``
             the duplicated, flipped and interlaced orblib
 
         """
         self.logger.debug('Checking for symmetric velocity array...')
         error_msg = 'velocity array must be symmetric'
-        assert np.allclose(orblib.xedg, -orblib.xedg[::-1]), error_msg
+        if orblib.y.ndim == 3:  # 1D histograms
+            assert np.allclose(orblib.xedg, -orblib.xedg[::-1]), error_msg
+        else:  # 2D histograms
+            assert np.allclose(orblib.xedg[0], -orblib.xedg[0][::-1]), error_msg
+            assert np.allclose(orblib.xedg[1], -orblib.xedg[1][::-1]), error_msg
         self.logger.debug('...check ok.')
-        losvd = orblib.y
-        n_orbs, n_vel_bins, n_spatial_bins = losvd.shape
-        reveresed_losvd = losvd[:, ::-1, :]
-        new_losvd = np.zeros((2*n_orbs, n_vel_bins, n_spatial_bins))
-        new_losvd[0::2] = losvd
-        new_losvd[1::2, :] = reveresed_losvd
-        new_orblib = dyn_kin.Histogram(xedg=orblib.xedg, y=new_losvd)
+        vel_d = orblib.y
+        if vel_d.ndim == 3:  # 1D histograms (losvd)
+            n_orbs, n_vel_bins, n_spatial_bins = vel_d.shape
+            reversed_vel_d = vel_d[:, ::-1, :]
+            new_vel_d = np.zeros((2*n_orbs, n_vel_bins, n_spatial_bins))
+        else:  # 2D histograms (proper motion)
+            n_orbs, n_vel_bins1, n_vel_bins2, n_spatial_bins = vel_d.shape
+            reversed_vel_d = vel_d[:, ::-1, ::-1, :]
+            new_vel_d = np.zeros((2*n_orbs,
+                                  n_vel_bins1,
+                                  n_vel_bins2,
+                                  n_spatial_bins))
+        new_vel_d[0::2] = vel_d
+        new_vel_d[1::2] = reversed_vel_d
+        if vel_d.ndim == 3:  # 1D histograms
+            new_orblib = dyn_kin.Histogram(xedg=orblib.xedg, y=new_vel_d)
+        else:  # 2D histograms
+            new_orblib = dyn_kin.Histogram2D(xedg=orblib.xedg, y=new_vel_d)
         return new_orblib
 
     def duplicate_flip_and_interlace_intmoms(self, intmom):
@@ -1037,54 +1168,80 @@ class LegacyOrbitLibrary(OrbitLibrary):
         return new_intmom
 
     def combine_orblibs(self, orblib1, orblib2):
-        """Combine two LOSVD histograms into one.
+        """Combine two velocity distribution histograms into one.
 
         Parameters
         ----------
-        orblib1 : ``dyn.kinematics.Histogram``
-        orblib2 : ``dyn.kinematics.Histogram``
+        orblib1 : ``dyn.kinematics.Histogram`` or ``dyn.kinematics.Histogram2D``
+        orblib2 : ``dyn.kinematics.Histogram`` or ``dyn.kinematics.Histogram2D``
 
         Returns
         -------
-        ``dyn.kinematics.Histogram``
+        ``dyn.kinematics.Histogram`` or ``dyn.kinematics.Histogram2D``
             the combined orbit libraries
 
         """
         # check orblibs are compatible
-        n_orbs1, n_vel_bins1, n_spatial_bins1 = orblib1.y.shape
-        n_orbs2, n_vel_bins2, n_spatial_bins2 = orblib2.y.shape
+        tmp1 = orblib1.y.shape
+        tmp2 = orblib2.y.shape
+        n_orbs1 = tmp1[0]
+        n_orbs2 = tmp2[0]
+        n_spatial_bins1 = tmp1[-1]
+        n_spatial_bins2 = tmp2[-1]
+        if orblib1.y.ndim == 3:  # 1D histograms
+            n_vel_bins1 = tmp1[1]
+            n_vel_bins2 = tmp2[1]
+        else:  # 2D histograms
+            n_vel_bins1 = (tmp1[1:3])
+            n_vel_bins2 = (tmp2[1:3])
         self.logger.debug('Checking number of velocity bins...')
         error_msg = 'orblibs have different number of velocity bins'
         assert n_vel_bins1==n_vel_bins2, error_msg
         self.logger.debug('Checking velocity arrays...')
         error_msg = 'orblibs have different velocity arrays'
-        assert np.array_equal(orblib1.x, orblib2.x), error_msg
+        if orblib1.y.ndim == 3:  # 1D histograms
+            assert np.array_equal(orblib1.x, orblib2.x), error_msg
+        else:  # 2D histograms
+            assert np.array_equal(orblib1.x[0], orblib2.x[0]), error_msg
+            assert np.array_equal(orblib1.x[1], orblib2.x[1]), error_msg
         self.logger.debug('Checking number of spatial bins...')
         error_msg = 'orblibs have different number of spatial bins'
         assert n_spatial_bins1==n_spatial_bins2, error_msg
         self.logger.debug('...checks ok.')
-        new_losvd = np.zeros((n_orbs1 + n_orbs2,
-                              n_vel_bins1,
-                              n_spatial_bins1))
-        new_losvd[:n_orbs1] = orblib1.y
-        new_losvd[n_orbs1:] = orblib2.y
-        new_orblib = dyn_kin.Histogram(xedg=orblib1.xedg, y=new_losvd)
+        if orblib1.y.ndim == 3:
+            new_vel_d = np.zeros((n_orbs1 + n_orbs2,
+                                  n_vel_bins1,
+                                  n_spatial_bins1))
+        else:
+            new_vel_d = np.zeros((n_orbs1 + n_orbs2,
+                                  n_vel_bins1[0],
+                                  n_vel_bins1[1],
+                                  n_spatial_bins1))
+        new_vel_d[:n_orbs1] = orblib1.y
+        new_vel_d[n_orbs1:] = orblib2.y
+        if orblib1.y.ndim == 3:
+            new_orblib = dyn_kin.Histogram(xedg=orblib1.xedg, y=new_vel_d)
+        else:
+            new_orblib = dyn_kin.Histogram2D(xedg=orblib1.xedg, y=new_vel_d)
         return new_orblib
 
-    def read_losvd_histograms(self, pops=False):
+    def read_vel_histograms(self, pops=False):
         """Read the orbit library
 
         Read box orbits and tube orbits, mirrors the latter, and combines.
-        Rescales the velocity axis according to the ``ml`` value. Sets LOSVDs
-        and 3D grid/aperture masses of the combined orbit library.
+        Rescales the velocity axis according to the ``ml`` value. Sets LOSVDs,
+        proper motion velocity distributions, and 3D grid/aperture masses of
+        the combined orbit library.
         If pops=True, only calculates the populations' projected masses.
 
         Returns
         -------
         If pops is False, sets the attributes:
-            -   ``self.losvd_histograms``: a list, whose i'th entry is a
-                ``dyn.kinematics.Histogram`` object holding the orbit lib LOSVDs
-                binned for the i'th kinematic set
+            -   ``self.vel_histograms``: a list, whose i'th entry is EITHER a
+                ``dyn.kinematics.Histogram`` object holding the orbit library
+                LOSVDs OR a ``dyn.kinematics.Histogram2D`` object holding the
+                orbit library proper motion velocity distributions, binned for
+                the i'th kinematic set
             -   ``self.intrinsic_masses``: 3D grid/intrinsic masses of orbit lib
             -   ``self.projected_masses``: aperture/proj. masses of orbit lib
             -   ``self.n_orbs``: number of orbits in the orbit library
@@ -1093,12 +1250,8 @@ class LegacyOrbitLibrary(OrbitLibrary):
                 populations data
 
         """
-        if self.system.is_bar_disk_system():
-            stars = self.system.get_unique_bar_component()
-        else:
-            stars = self.system.get_unique_triaxial_visible_component()
-        n_kins = len(stars.kinematic_data)
-        n_pops = len(stars.population_data) if pops else 0
+        n_kins = len(self.stars.kinematic_data)
+        n_pops = len(self.stars.population_data) if pops else 0
 
         # TODO: check if this ordering is compatible with weights read in by
         # LegacyWeightSolver.read_weights
@@ -1141,12 +1294,17 @@ class LegacyOrbitLibrary(OrbitLibrary):
             density_3D = np.vstack((tube_density_3D, box_density_3D))
             for i in range(n_kins):
                 orblib[i].scale_x_values(self.velocity_scaling_factor)
-            self.losvd_histograms = orblib
+            self.vel_histograms = orblib
             self.intrinsic_masses = density_3D
             self.n_orbs = orblib[0].y.shape[0] if n_kins > 0 else 0
-            proj_mass = [np.sum(orblib[i].y,1) for i in range(n_kins)]
+            proj_mass = []
+            for kin in range(n_kins):
+                if self.vel_histograms[kin].y.ndim == 3:  # 1D histograms
+                    proj_mass.append(np.sum(self.vel_histograms[kin].y, 1))
+                else:  # 2D histograms
+                    proj_mass.append(np.sum(self.vel_histograms[kin].y, (1,2)))
             self.projected_masses = proj_mass
-        else:
+        else:  # note: pops cannot share apertures with 2d histogram kins
             proj_mass = [np.sum(orblib[i].y,1) for i in range(n_pops)]
             self.pops_projected_masses = proj_mass
 
@@ -1437,7 +1595,7 @@ class LegacyOrbitLibrary(OrbitLibrary):
         n_ytish = np.sum(bool_ytish)
         n_ztish = np.sum(bool_ztish)
         n_other = np.sum(bool_other)
-        # n_sum = n_box + n_xtish + n_ytish + n_ztish + n_other
+        # UNUSED n_sum = n_box + n_xtish + n_ytish + n_ztish + n_other
         n_xt_exact = np.sum(bool_xtube)
         n_yt_exact = np.sum(bool_ytube)
         n_zt_exact = np.sum(bool_ztube)