+++ dynamite/weight_solvers.py
@@ -14,7 +14,6 @@ except ModuleNotFoundError:
 
 from dynamite import constants
 from dynamite import analysis
-from dynamite import physical_system as physys
 from dynamite import kinematics as dyn_kin
 
 
@@ -107,10 +106,9 @@ class WeightSolver(object):
             squared residuals of V, sigma, and GH coefficients from h_3 to h_N
 
         """
-        stars = \
-          self.system.get_component_from_class(physys.TriaxialVisibleComponent)
+        stars = self.system.get_unique_triaxial_visible_component()
         if any(k.type != 'GaussHermite' for k in stars.kinematic_data):
-            self.logger.info("'GaussHermite' kinematics required for "
+            self.logger.info("All kinematics must be 'GaussHermite' for "
                              "kinmapchi2. Value set to nan.")
             return float('nan')  # #######################################
         number_gh = self.settings['number_GH']
@@ -642,6 +640,8 @@ class NNLS(WeightSolver):
                 - ``self.intrinsic_mass_error``
                 - ``self.projected_masses``
                 - ``self.projected_mass_error``
+                - ``self.total_mass``
+                - ``self.total_mass_error``
                 -   constraint counts ``self.n_intrinsic``, ``self.n_apertures``
                     and ``self.n_mass_constraints``
 
@@ -688,7 +688,7 @@ class NNLS(WeightSolver):
 
         """
         # construct vector of observed constraints (con), errors (econ) and
-        # matrix or orbit properties (orbmat)
+        # matrix of orbit propertites (orbmat)
         con = np.zeros(self.n_mass_constraints)
         econ = np.zeros(self.n_mass_constraints)
         orbmat = np.zeros((self.n_mass_constraints, orblib.n_orbs))
@@ -715,11 +715,12 @@ class NNLS(WeightSolver):
         orbmat[idx,:] = np.hstack(orblib.projected_masses).T
         # add kinematics to con, econ, orbmat
         stars = self.system.get_unique_triaxial_visible_component()
-        kins_and_orb_losvds = zip(stars.kinematic_data, orblib.losvd_histograms)
+        kins_and_orb_veldist = zip(stars.kinematic_data, orblib.vel_histograms)
         idx_ap_start = 0
-        for (kins, orb_losvd) in kins_and_orb_losvds:
+        for (kins, orb_veldist) in kins_and_orb_veldist:
+            hist_dim = len(orb_veldist.y[0,...,0].shape)  # 1D or 2D vel hists
             # pick out the projected masses for this kinematic set
-            n_ap = kins.n_spatial_bins  # OK for both GaussHermite & BayesLOSVD
+            n_ap = kins.n_spatial_bins  # OK for all kinematics
             idx_ap_end = idx_ap_start + n_ap
             prj_mass_i = self.projected_masses[idx_ap_start:idx_ap_end]
             idx_ap_start += n_ap
@@ -728,16 +729,17 @@ class NNLS(WeightSolver):
             obs_kins, obs_kins_err = tmp
             obs_kins = (obs_kins.T * prj_mass_i).T
             obs_kins_err = (obs_kins_err.T * prj_mass_i).T
-            # set the first and last point in the velocity histograms to zero
-            # to mimic what is done in `triaxnnnls_CRcut.f90`
-            orb_losvd.y[:,0,:] = 0.
-            orb_losvd.y[:,-1,:] = 0.
+            if hist_dim == 1:  # Do we need this for proper motions (2d hists)?
+                # set the first and last point in the velocity histograms to
+                # zero to mimic what is done in `triaxnnnls_CRcut.f90`
+                orb_veldist.y[:,0,:] = 0.
+                orb_veldist.y[:,-1,:] = 0.
             # transform orblib to same parameterisation as observed kinematics
-            orb_kins = kins.transform_orblib_to_observables(orb_losvd,
+            orb_kins = kins.transform_orblib_to_observables(orb_veldist,
                                                             self.settings)
             if self.CRcut:
                 # note: this only has an effect if type(kins) is GaussHermite
-                orb_kins = self.apply_CR_cut(kins, orb_losvd, orb_kins)
+                orb_kins = self.apply_CR_cut(kins, orb_veldist, orb_kins)
             # append constraints/errors/orbits to con/econ/orbmat
             obs_kins = np.ravel(obs_kins)
             con = np.concatenate((con, obs_kins))
@@ -746,9 +748,25 @@ class NNLS(WeightSolver):
             orb_kins = np.reshape(orb_kins, (orblib.n_orbs, -1))
             orbmat = np.vstack((orbmat, orb_kins.T))
         # divide constraint vector and matrix by errors
-        rhs = con/econ
-        orbmat = (orbmat.T/econ).T
-        return orbmat, rhs
+        if np.any(con[econ==0] != 0):
+            txt = 'Weight solving fail: zero errors for nonzero constraints!'
+            self.logger.error(txt)
+            raise ValueError(txt)
+        # previous statement: rhs = con/econ, np.divide has the "where" clause
+        rhs = np.zeros_like(con)
+        np.divide(con, econ, out=rhs, where=econ!=0)  # con = econ = 0 is ok
+        if np.any(np.ravel(orbmat[econ==0]) != 0):
+            err_loc = np.nonzero(((orbmat != 0).T * (econ == 0)).T)
+            txt = f'Weight solving problem in {self.direc_with_ml}: ' \
+                  'zero errors for nonzero matrix coefficients at ' \
+                  f'[constraint no, orbit no] = {err_loc}! Matrix value(s) ' \
+                  f'there ({orbmat[err_loc]}) will be considered zero.'
+            self.logger.warning(txt)
+            orbmat[err_loc] = 0
+        # previous statement: orbmat = (orbmat.T/econ).T, np.divide has "where"
+        orbmat = orbmat.T
+        np.divide(orbmat, econ, out=orbmat, where=econ!=0)
+        return orbmat.T, rhs
 
     def apply_CR_cut(self, kins, orb_losvd, orb_gh):
         r"""apply `CRcut`
@@ -801,10 +819,14 @@ class NNLS(WeightSolver):
         **Note:** the returned chi2 values are not the same as
         ``LegacyWeightSolver.read_chi2`` - see the docstring for more info
 
+        Apart from weight solving, the attributes ``orblib.intrinsic_masses``,
+        ``orblib.projected_masses``, and ``orblib.vel_histograms`` are set
+        via calling ``orblib.read_vel_histograms()``.
+
         Parameters
         ----------
         orblib : dyn.OrbitLibrary
-        ignore_existing_weights : bool
+        ignore_existing_weights : bool, optional
             If True, do not check for already existing weights and solve again.
             Default is False.
 
@@ -832,9 +854,9 @@ class NNLS(WeightSolver):
             chi2_kin = results.meta['chi2_kin']
             chi2_kinmap = results.meta['chi2_kinmap']
         else:
-            orblib.read_losvd_histograms()  # sets orblib.losvd_histograms,
-                                            # orblib.intrinsic_masses, and
-                                            # orblib.projected_masses
+            orblib.read_vel_histograms()  # sets orblib.vel_histograms,
+                                          # orblib.intrinsic_masses, and
+                                          # orblib.projected_masses
             A, b = self.construct_nnls_matrix_and_rhs(orblib)
             if self.nnls_solver=='scipy':
                 try: